feat: Board and History work.
The board recognizes when the game has already ended and stops working
when that happens.

Besides, history buttons work correctly, since now not only the board state
is stored, but the correspondent turn associated with the move.

Although the winner is already being calculated, it is not yet being
displayed in a label. That's the last thing to accomplish for the next
commit and for the project to be concluded.